+++ CHANGELOG.md
@@ -8,6 +8,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- `Fruit Kick` and `Plucked!` plugin data parsers.
+
+### Changed
+
+- Unbound descriptors return `self` - more `property`*esque* behaviour.
+  This is primarily done to allow flpinspect to inspect descriptor types.
+
 ## [2.0.0] - 2023-03-18
 
 Welcome PyFLP 2.0 🎉